+++ api/lib/infrastructure/repositories/target-profile-share-repository.js
@@ -1,16 +1,6 @@
 import { knex } from '../../../db/knex-database-connection.js';
 import { DomainTransaction } from '../DomainTransaction.js';
 
-const addTargetProfilesToOrganization = async function ({ organizationId, targetProfileIdList }) {
-  const targetProfileShareToAdd = targetProfileIdList.map((targetProfileId) => {
-    return { organizationId, targetProfileId };
-  });
-  await knex('target-profile-shares')
-    .insert(targetProfileShareToAdd)
-    .onConflict(['targetProfileId', 'organizationId'])
-    .ignore();
-};
-
 const batchAddTargetProfilesToOrganization = async function (
   organizationTargetProfiles,
   domainTransaction = DomainTransaction.emptyTransaction(),
@@ -20,4 +10,4 @@ const batchAddTargetProfilesToOrganization = async function (
     .transacting(domainTransaction.knexTransaction);
 };
 
-export { addTargetProfilesToOrganization, batchAddTargetProfilesToOrganization };
+export { batchAddTargetProfilesToOrganization };
