+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -1149,75 +1149,6 @@ describe('Acceptance | Application | organization-controller', function () {
     });
   });
 
-  describe('POST /api/admin/organizations/{id}/attach-target-profiles', function () {
-    let userId;
-    let organizationId;
-    let alreadyAttachedTargetProfileId;
-    let toAttachTargetProfileId;
-
-    beforeEach(async function () {
-      userId = databaseBuilder.factory.buildUser.withRole().id;
-      organizationId = databaseBuilder.factory.buildOrganization().id;
-      alreadyAttachedTargetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-      toAttachTargetProfileId = databaseBuilder.factory.buildTargetProfile().id;
-      databaseBuilder.factory.buildTargetProfileShare({
-        organizationId,
-        targetProfileId: alreadyAttachedTargetProfileId,
-      });
-      await databaseBuilder.commit();
-    });
-
-    context('when target profiles to attach exists', function () {
-      it('should attach target profiles to organization', async function () {
-        // given
-        const options = {
-          method: 'POST',
-          url: `/api/admin/organizations/${organizationId}/attach-target-profiles`,
-          headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-          payload: {
-            'target-profile-ids': [alreadyAttachedTargetProfileId, toAttachTargetProfileId],
-          },
-        };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        const attachedTargetProfileIds = await knex('target-profile-shares')
-          .pluck('targetProfileId')
-          .where({ organizationId })
-          .orderBy('targetProfileId', 'ASC');
-        expect(response.statusCode).to.equal(204);
-        expect(attachedTargetProfileIds).to.deepEqualArray([alreadyAttachedTargetProfileId, toAttachTargetProfileId]);
-      });
-    });
-
-    context('when a target profile does not exist', function () {
-      it('should return a 404 error without attaching any target profile', async function () {
-        // given
-        const options = {
-          method: 'POST',
-          url: `/api/admin/organizations/${organizationId}/attach-target-profiles`,
-          headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-          payload: {
-            'target-profile-ids': [alreadyAttachedTargetProfileId, 6000, toAttachTargetProfileId],
-          },
-        };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        const attachedTargetProfileIds = await knex('target-profile-shares')
-          .pluck('targetProfileId')
-          .where({ organizationId })
-          .orderBy('targetProfileId', 'ASC');
-        expect(response.statusCode).to.equal(404);
-        expect(attachedTargetProfileIds).to.deepEqualArray([alreadyAttachedTargetProfileId]);
-      });
-    });
-  });
-
   describe('GET /api/organizations/{organizationId}/divisions', function () {
     it('should return the divisions', async function () {
       // given
