+++ api/tests/prescription/target-profile/integration/infrastructure/repositories/organizations-to-attach-to-target-profile-repository_test.js
@@ -1,8 +1,88 @@
+import _ from 'lodash';
+
 import * as organizationsToAttachToTargetProfileRepository from '../../../../../../src/prescription/target-profile/infrastructure/repositories/organizations-to-attach-to-target-profile-repository.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, knex } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | Organizations-to-attach-to-target-profile', function () {
+  describe('#addTargetProfilesToOrganization', function () {
+    let organizationId;
+    let targetProfileIdA;
+    let targetProfileIdB;
+    let targetProfileIdC;
+
+    beforeEach(function () {
+      organizationId = databaseBuilder.factory.buildOrganization().id;
+      targetProfileIdA = databaseBuilder.factory.buildTargetProfile().id;
+      targetProfileIdB = databaseBuilder.factory.buildTargetProfile().id;
+      targetProfileIdC = databaseBuilder.factory.buildTargetProfile().id;
+      return databaseBuilder.commit();
+    });
+
+    it('should save all the target profile shares for the organization', async function () {
+      // given
+      const targetProfileIdList = [targetProfileIdA, targetProfileIdB, targetProfileIdC];
+
+      // when
+      await organizationsToAttachToTargetProfileRepository.addTargetProfilesToOrganization({
+        organizationId,
+        targetProfileIdList,
+      });
+
+      // then
+      const targetProfileShares = await knex('target-profile-shares').where({ organizationId });
+      expect(targetProfileShares).to.have.lengthOf(3);
+      expect(_.map(targetProfileShares, 'targetProfileId')).exactlyContain([
+        targetProfileIdA,
+        targetProfileIdB,
+        targetProfileIdC,
+      ]);
+    });
+
+    it('should not create another target profile share nor throw an error when it already exists', async function () {
+      // given
+      databaseBuilder.factory.buildTargetProfileShare({ organizationId, targetProfileId: targetProfileIdA });
+      const targetProfileIdList = [targetProfileIdA, targetProfileIdB, targetProfileIdC];
+
+      // when
+      await organizationsToAttachToTargetProfileRepository.addTargetProfilesToOrganization({
+        organizationId,
+        targetProfileIdList,
+      });
+
+      // then
+      const targetProfileShares = await knex('target-profile-shares').where({ organizationId });
+      expect(targetProfileShares).to.have.lengthOf(3);
+      expect(_.map(targetProfileShares, 'targetProfileId')).to.have.members([
+        targetProfileIdA,
+        targetProfileIdB,
+        targetProfileIdC,
+      ]);
+    });
+
+    it('should not erase old target profil share', async function () {
+      // given
+      databaseBuilder.factory.buildTargetProfileShare({ organizationId, targetProfileId: targetProfileIdA });
+      await databaseBuilder.commit();
+      const targetProfileIdList = [targetProfileIdB, targetProfileIdC];
+
+      // when
+      await organizationsToAttachToTargetProfileRepository.addTargetProfilesToOrganization({
+        organizationId,
+        targetProfileIdList,
+      });
+
+      // then
+      const targetProfileShares = await knex('target-profile-shares').where({ organizationId });
+      expect(targetProfileShares).to.have.lengthOf(3);
+      expect(_.map(targetProfileShares, 'targetProfileId')).exactlyContain([
+        targetProfileIdA,
+        targetProfileIdB,
+        targetProfileIdC,
+      ]);
+    });
+  });
+
   describe('#attachOrganizations', function () {
     it('should return attachedIds', async function () {
       const targetProfileId = databaseBuilder.factory.buildTargetProfile().id;
