+++ api/tests/unit/application/organizations/index_test.js
@@ -139,169 +139,6 @@ describe('Unit | Router | organization-router', function () {
     });
   });
 
-  describe('POST /api/admin/organizations/{id}/attach-target-profiles', function () {
-    it('should allow to controller if user has role SUPER_ADMIN', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': [1, 2],
-      };
-
-      // when
-      await httpTestServer.request('POST', '/api/admin/organizations/1/attach-target-profiles', payload);
-
-      // then
-      sinon.assert.calledOnce(organizationController.attachTargetProfiles);
-    });
-
-    it('should allow to controller if user has role SUPPORT', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': [1, 2],
-      };
-
-      // when
-      await httpTestServer.request('POST', '/api/admin/organizations/1/attach-target-profiles', payload);
-
-      // then
-      sinon.assert.calledOnce(organizationController.attachTargetProfiles);
-    });
-
-    it('should allow to controller if user has role METIER', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': [1, 2],
-      };
-
-      // when
-      await httpTestServer.request('POST', '/api/admin/organizations/1/attach-target-profiles', payload);
-
-      // then
-      sinon.assert.calledOnce(organizationController.attachTargetProfiles);
-    });
-
-    it('should return 403 without reaching controller if user has not an allowed role', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': [1, 2],
-      };
-
-      // when
-      const response = await httpTestServer.request(
-        'POST',
-        '/api/admin/organizations/1/attach-target-profiles',
-        payload,
-      );
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      sinon.assert.notCalled(organizationController.attachTargetProfiles);
-    });
-
-    it('should return a 404 HTTP response when target-profile-ids do not contain only numbers', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': ['a', 2],
-      };
-
-      // when
-      const response = await httpTestServer.request(
-        'POST',
-        '/api/admin/organizations/1/attach-target-profiles',
-        payload,
-      );
-
-      // then
-      expect(response.statusCode).to.equal(404);
-      expect(response.payload).to.have.string("L'id d'un des profils cible ou de l'organisation n'est pas valide");
-    });
-
-    it('should return a 404 HTTP response when organization id is not valid', async function () {
-      // given
-      sinon.stub(organizationController, 'attachTargetProfiles').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const payload = {
-        'target-profile-ids': [1, 2],
-      };
-
-      // when
-      const response = await httpTestServer.request(
-        'POST',
-        '/api/admin/organizations/coucou/attach-target-profiles',
-        payload,
-      );
-
-      // then
-      expect(response.statusCode).to.equal(404);
-      expect(response.payload).to.have.string("L'id d'un des profils cible ou de l'organisation n'est pas valide");
-    });
-  });
-
   describe('GET /api/admin/organizations/{id}/target-profile-summaries', function () {
     it('should allow to controller if user has role SUPER_ADMIN', async function () {
       // given
